+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Refactor Request and ConsolePanel components for improved UI and functionality. (`4adfc0e`)
 - Enhance UI components with resizable features and code cleanup. (`c735424`)
 - Refactor application structure and enhance request handling UI. (`a7b8bff`)
 - Refactor UI component imports and remove unused classes. (`af56f8a`)